+++ common/persistence/sql-extensions/mysql/visibility.go
